To-Do Schema
title: a required string that must be provided when creating a To-Do item.
description: an optional string that allows a user to add details about the To-Do.
completed: a boolean indicating whether the To-Do is completed, defaulted to false.
createdAt: a date indicating when the To-Do was created, defaulted to the current date and time.
dueDate: an optional date specifying when the To-Do should be completed.